# docs/swagger.yaml
definitions:
...
  controllers.ChainResult:
    properties:
      current_page:
        description: 当前页数
        type: integer
      page_data:
        description: 页的数据
        items:
          $ref: '#/definitions/dao.ChainInfo'
        type: array
      page_size:
        description: 页的大小
        type: integer
      total_count:
        description: 总记录数
        type: integer
    type: object
...
  controllers.JsonResult:
    properties:
      code:
        type: integer
      data:
        type: object
      err_msg:
        type: string
    required:
    - code
    type: object
...
  dao.Chain:
    properties:
      coin_name:
        description: 币名
        type: string
      contractInstanceId:
        description: 合约实例
        type: integer
      name:
        description: 链的名称
        type: string
      network_id:
        description: 链的网络编号
        type: integer
      symbol:
        description: 符号
        type: string
    required:
    - coin_name
    - name
    - network_id
    - symbol
    type: object
  dao.ChainInfo:
    properties:
      address:
        description: 合约地址
        type: string
      coin_name:
        description: 币名
        type: string
      contract_instance_id:
        description: 合约实例
        type: integer
      name:
        description: 链的名称
        type: string
      network_id:
        description: 链的网络编号
        type: integer
      symbol:
        description: 符号
        type: string
    type: object
...
  dao.Node:
    properties:
      address:
        description: 地址
        type: string
      chain_id:
        description: 链id
        type: integer
      is_https:
        type: boolean
      name:
        type: string
      port:
        description: 端口
        type: integer
      user_id:
        type: integer
    required:
    - address
    - chain_id
    - name
    - port
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /chain/{chain_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: chainId
        in: path
        name: chain_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除链信息
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除链信息
      tags:
      - chain
  /chain/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 链的名称
        in: formData
        name: name
        required: true
        type: integer
      - description: 链的网络编号
        in: formData
        name: network_id
        required: true
        type: string
      - description: 币的名称
        in: formData
        name: coin_name
        required: true
        type: integer
      - description: 币的符号
        in: formData
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 添加链信息
      tags:
      - node
...
  /chain/info/{chain_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 链id
        in: path
        name: chain_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dao.ChainInfo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取链信息
      tags:
      - GetChainInfo
  /chain/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 当前页，默认1
        in: query
        name: current_page
        required: true
        type: string
      - description: 页的记录数，默认10
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/controllers.ChainResult'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 链的管理
      tags:
      - ListChain
  /chain/list/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dao.Chain'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有链信息
      tags:
      - ListAllChain
  /chain/node:
    get:
      consumes:
      - application/json
      parameters:
      - description: 链id
        in: query
        name: chain_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dao.Node'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取链相关的节点
      tags:
      - GetNodeByChain
...
  /chain/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: 链的id
        in: formData
        name: id
        required: true
        type: integer
      - description: 链的名称
        in: formData
        name: name
        required: true
        type: integer
      - description: 链的网络编号
        in: formData
        name: network_id
        required: true
        type: string
      - description: 币的名称
        in: formData
        name: coin_name
        required: true
        type: integer
      - description: 币的符号
        in: formData
        name: symbol
        required: true
        type: string
      - description: 合约实例id
        in: formData
        name: contract_instance_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.JsonResult'
            - properties:
                data:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 编辑链信息
      tags:
      - node
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"